+++ model/iterator.go
@@ -1,4 +1,4 @@
-package storage
+package model
 
 import "context"
