+++ model/test_test.go
@@ -1,197 +1,288 @@
 package model
 
 import (
+	"context"
+	"io"
 	"strings"
 	"testing"
 	"time"
 
-	"github.com/evergreen-ci/logkeeper/db"
+	"github.com/evergreen-ci/logkeeper/env"
 	"github.com/evergreen-ci/logkeeper/testutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/mgo.v2/bson"
 )
 
-func TestIncrementTestSequence(t *testing.T) {
-	require.NoError(t, testutil.InitDB())
-	require.NoError(t, testutil.ClearCollections(TestsCollection))
-
-	testID := NewTestID(time.Time{})
-	test := &Test{Id: testID, Seq: 1}
-	require.NoError(t, test.Insert())
-
-	assert.NoError(t, test.IncrementSequence(1))
-	assert.Equal(t, 2, test.Seq)
-
-	test, err := FindTestByID(string(testID))
-	assert.NoError(t, err)
-	assert.Equal(t, test.Seq, 2)
-}
-
-func TestFindTestsForBuild(t *testing.T) {
-	require.NoError(t, testutil.InitDB())
-	require.NoError(t, testutil.ClearCollections(TestsCollection))
-
-	require.NoError(t, (&Test{Id: NewTestID(time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)), Name: "t0", BuildId: "b0", Started: time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)}).Insert())
-	require.NoError(t, (&Test{Id: NewTestID(time.Date(2009, time.November, 10, 23, 1, 0, 0, time.UTC)), Name: "t1", BuildId: "b0", Started: time.Date(2009, time.November, 10, 23, 1, 0, 0, time.UTC)}).Insert())
-	require.NoError(t, (&Test{Id: NewTestID(time.Time{}), BuildId: "b1"}).Insert())
-
-	tests, err := FindTestsForBuild("b0")
-	assert.NoError(t, err)
-	require.Len(t, tests, 2)
-	assert.Equal(t, tests[0].Name, "t0")
-	assert.Equal(t, tests[1].Name, "t1")
-}
-
-func TestRemoveTestsForBuild(t *testing.T) {
-	require.NoError(t, testutil.InitDB())
-	require.NoError(t, testutil.ClearCollections(TestsCollection))
-
-	require.NoError(t, (&Test{Id: NewTestID(time.Time{}), BuildId: "b0"}).Insert())
-	require.NoError(t, (&Test{Id: NewTestID(time.Time{}), BuildId: "b0"}).Insert())
-	require.NoError(t, (&Test{Id: NewTestID(time.Time{}), BuildId: "b1"}).Insert())
-
-	count, err := RemoveTestsForBuild("b0")
-	assert.NoError(t, err)
-	assert.Equal(t, count, 2)
-}
-
-func TestFindNext(t *testing.T) {
-	require.NoError(t, testutil.InitDB())
-	require.NoError(t, testutil.ClearCollections(TestsCollection))
-
-	t0 := Test{Id: NewTestID(time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)), Name: "t0", BuildId: "b0", Started: time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)}
-	t1 := Test{Id: NewTestID(time.Date(2009, time.November, 10, 23, 1, 0, 0, time.UTC)), Name: "t1", BuildId: "b0", Started: time.Date(2009, time.November, 10, 23, 1, 0, 0, time.UTC)}
-	require.NoError(t, t0.Insert())
-	require.NoError(t, t1.Insert())
-
-	next, err := t0.findNext()
-	assert.NoError(t, err)
-	assert.Equal(t, t1.Name, next.Name)
-}
-
-func TestGetExecutionWindow(t *testing.T) {
-	require.NoError(t, testutil.InitDB())
-
-	t.Run("NoLaterTest", func(t *testing.T) {
-		require.NoError(t, testutil.ClearCollections(TestsCollection))
-
-		startTime := time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)
-		t0 := Test{Id: NewTestID(startTime), Name: "t0", BuildId: "b0", Started: startTime}
-		assert.NoError(t, t0.Insert())
-		minTime, maxTime, err := t0.GetExecutionWindow()
-		assert.NoError(t, err)
-		assert.True(t, t0.Started.Equal(minTime))
-		assert.Nil(t, maxTime)
-	})
-	t.Run("LaterTest", func(t *testing.T) {
-		require.NoError(t, testutil.ClearCollections(TestsCollection))
-
-		startTime := time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)
-		t0 := Test{Id: NewTestID(startTime), Name: "t0", BuildId: "b0", Started: time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)}
-		assert.NoError(t, t0.Insert())
-		t1 := Test{Id: NewTestID(startTime.Add(time.Hour)), Name: "t1", BuildId: "b0", Started: startTime.Add(time.Hour)}
-		assert.NoError(t, t1.Insert())
-		minTime, maxTime, err := t0.GetExecutionWindow()
-		assert.NoError(t, err)
-		assert.True(t, t0.Started.Equal(minTime))
-		require.NotNil(t, maxTime)
-		assert.True(t, t1.Started.Equal(*maxTime))
-	})
-}
-
 func TestNewTestID(t *testing.T) {
 	assert.True(t, strings.HasPrefix(string(NewTestID(time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC))), "1174efedab186000"))
 }
 
 func TestTestID(t *testing.T) {
-	require.NoError(t, testutil.InitDB())
-
-	t.Run("Timestamp", func(t *testing.T) {
-		t.Run("ObjectID", func(t *testing.T) {
-			startTime := time.Date(2009, time.November, 10, 23, 0, 0, 1, time.UTC)
-			objectID := bson.NewObjectIdWithTime(startTime)
-			newID := TestID(objectID.Hex())
-			assert.True(t, startTime.Equal(newID.Timestamp().Add(time.Nanosecond)))
-		})
-
-		t.Run("TestID", func(t *testing.T) {
-			startTime := time.Date(2009, time.November, 10, 23, 0, 0, 1, time.UTC)
-			newID := NewTestID(startTime)
-			assert.True(t, startTime.Equal(newID.Timestamp()))
-		})
+	t.Run("ObjectID", func(t *testing.T) {
+		startTime := time.Date(2009, time.November, 10, 23, 0, 0, 1, time.UTC)
+		objectID := bson.NewObjectIdWithTime(startTime)
+		newID := objectID.Hex()
+		assert.True(t, startTime.Equal(testIDTimestamp(newID).Add(time.Nanosecond)))
 	})
-	t.Run("SetBSON", func(t *testing.T) {
-		db, closer := db.DB()
-		defer closer()
 
-		t.Run("ObjectIDUnmarshalledToTestID", func(t *testing.T) {
-			require.NoError(t, testutil.ClearCollections(TestsCollection))
-			objectID := bson.NewObjectId()
-			require.NoError(t, db.C(TestsCollection).Insert(bson.M{"_id": objectID}))
-
-			var test Test
-			assert.NoError(t, db.C(TestsCollection).Find(bson.M{}).One(&test))
-			assert.Equal(t, TestID(objectID.Hex()), test.Id)
-		})
-
-		t.Run("TestIDUnmarshalledToTestID", func(t *testing.T) {
-			require.NoError(t, testutil.ClearCollections(TestsCollection))
-
-			newID := NewTestID(time.Time{})
-			require.NoError(t, db.C(TestsCollection).Insert(bson.M{"_id": string(newID)}))
-
-			var test Test
-			assert.NoError(t, db.C(TestsCollection).Find(bson.M{}).One(&test))
-			assert.Equal(t, newID, test.Id)
-		})
+	t.Run("TestID", func(t *testing.T) {
+		startTime := time.Date(2009, time.November, 10, 23, 0, 0, 1, time.UTC)
+		newID := NewTestID(startTime)
+		assert.True(t, startTime.Equal(testIDTimestamp(newID)))
 	})
-	t.Run("GetBSON", func(t *testing.T) {
-		db, closer := db.DB()
-		defer closer()
+}
 
-		t.Run("ObjectIDInsertedAsObjectID", func(t *testing.T) {
-			require.NoError(t, testutil.ClearCollections(TestsCollection))
-			objectID := bson.NewObjectId()
-			require.NoError(t, db.C(TestsCollection).Insert(Test{Id: TestID(objectID.Hex())}))
+func TestUploadTestMetadata(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
 
-			var testDoc struct {
-				ID bson.ObjectId `bson:"_id"`
+	defer testutil.SetBucket(t, "")()
+	test := Test{
+		ID:      "62dba0159041307f697e6ccc",
+		Name:    "test0",
+		BuildID: "5a75f537726934e4b62833ab6d5dca41",
+		TaskID:  "t0",
+		Phase:   "phase0",
+		Command: "command0",
+	}
+	expectedData, err := test.toJSON()
+	require.NoError(t, err)
+	require.NoError(t, test.UploadTestMetadata(ctx))
+
+	r, err := env.Bucket().Get(ctx, "/builds/5a75f537726934e4b62833ab6d5dca41/tests/62dba0159041307f697e6ccc/metadata.json")
+	require.NoError(t, err)
+	data, err := io.ReadAll(r)
+	require.NoError(t, err)
+	assert.Equal(t, expectedData, data)
+}
+
+func TestTestKey(t *testing.T) {
+	test := Test{
+		ID:      "test0",
+		Name:    "name",
+		BuildID: "build0",
+		TaskID:  "t0",
+		Phase:   "phase0",
+		Command: "command0",
+	}
+	assert.Equal(t, "builds/build0/tests/test0/metadata.json", test.key())
+}
+
+func TestTestToJSON(t *testing.T) {
+	test := Test{
+		ID:      "test0",
+		Name:    "name",
+		BuildID: "build0",
+		TaskID:  "t0",
+		Phase:   "phase0",
+		Command: "command0",
+	}
+	data, err := test.toJSON()
+	require.NoError(t, err)
+	assert.JSONEq(t, `{"id":"test0","name":"name","build_id":"build0","task_id":"t0","phase":"phase0","command":"command0"}`, string(data))
+}
+
+func TestCheckTestMetadata(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	defer testutil.SetBucket(t, "../testdata/simple")()
+
+	for _, test := range []struct {
+		name     string
+		buildID  string
+		testID   string
+		expected bool
+	}{
+		{
+			name:     "BuildExists",
+			buildID:  "5a75f537726934e4b62833ab6d5dca41",
+			expected: true,
+		},
+		{
+			name:     "TestExists",
+			buildID:  "5a75f537726934e4b62833ab6d5dca41",
+			testID:   "17046404de18d0000000000000000000",
+			expected: true,
+		},
+		{
+			name:     "BuildDNE",
+			buildID:  "DNE",
+			expected: false,
+		},
+		{
+			name:     "BuildExistsTestDNE",
+			buildID:  "5a75f537726934e4b62833ab6d5dca41",
+			testID:   "DNE",
+			expected: false,
+		},
+		{
+			name:     "BuildDNETestDNE",
+			buildID:  "DNE",
+			testID:   "DNE",
+			expected: false,
+		},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			actual, err := CheckTestMetadata(ctx, test.buildID, test.testID)
+			require.NoError(t, err)
+			assert.Equal(t, test.expected, actual)
+		})
+	}
+}
+
+func TestFindTestByID(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	defer testutil.SetBucket(t, "../testdata/simple")()
+	t.Run("Exists", func(t *testing.T) {
+		expected := &Test{
+			ID:      "17046404de18d0000000000000000000",
+			BuildID: "5a75f537726934e4b62833ab6d5dca41",
+			Name:    "geo_max:CheckReplOplogs",
+			TaskID:  "mongodb_mongo_master_enterprise_rhel_80_64_bit_multiversion_all_feature_flags_retryable_writes_downgrade_last_continuous_2_enterprise_f98b3361fbab4e02683325cc0e6ebaa69d6af1df_22_07_22_11_24_37",
+			Phase:   "phase0",
+			Command: "command0",
+		}
+		actual, err := FindTestByID(ctx, "5a75f537726934e4b62833ab6d5dca41", "17046404de18d0000000000000000000")
+		require.NoError(t, err)
+		assert.Equal(t, expected, actual)
+	})
+	t.Run("BuildDNE", func(t *testing.T) {
+		test, err := FindTestByID(ctx, "DNE", "17046404de18d0000000000000000000")
+		require.NoError(t, err)
+		assert.Nil(t, test)
+	})
+	t.Run("TestDNE", func(t *testing.T) {
+		test, err := FindTestByID(ctx, "5a75f537726934e4b62833ab6d5dca41", "DNE")
+		require.NoError(t, err)
+		assert.Nil(t, test)
+	})
+}
+
+func TestFindTestsForBuild(t *testing.T) {
+	defer testutil.SetBucket(t, "../testdata/between")()
+
+	expected := []Test{
+		{
+			ID:      "0de0b6b3bf4ac6400000000000000000",
+			BuildID: "5a75f537726934e4b62833ab6d5dca41",
+			Name:    "geo_max:CheckReplOplogs",
+			TaskID:  "Task",
+			Command: "command0",
+			Phase:   "phase0",
+		},
+		{
+			ID:      "0de0b6b3cb3688400000000000000000",
+			BuildID: "5a75f537726934e4b62833ab6d5dca41",
+			Name:    "geo_max:CheckReplOplogs2",
+			TaskID:  "Task",
+			Command: "command1",
+			Phase:   "phase1",
+		},
+	}
+	testResponse, err := FindTestsForBuild(context.Background(), "5a75f537726934e4b62833ab6d5dca41")
+	require.NoError(t, err)
+	assert.Equal(t, expected, testResponse)
+}
+
+func TestTestExecutionWindow(t *testing.T) {
+	t.Run("NoLaterTest", func(t *testing.T) {
+		startTime := time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)
+		allTestIDs := []string{
+			NewTestID(startTime),
+		}
+		tr, err := testExecutionWindow(allTestIDs, string(allTestIDs[0]))
+		assert.NoError(t, err)
+		assert.True(t, tr.StartAt.Equal(startTime))
+		assert.True(t, tr.EndAt.Equal(TimeRangeMax))
+	})
+
+	t.Run("LaterTest", func(t *testing.T) {
+		startTime := time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)
+		allTestIDs := []string{
+			NewTestID(startTime),
+			NewTestID(startTime.Add(time.Hour)),
+		}
+		tr, err := testExecutionWindow(allTestIDs, string(allTestIDs[0]))
+		assert.NoError(t, err)
+		assert.True(t, tr.StartAt.Equal(startTime))
+		assert.True(t, tr.EndAt.Equal(startTime.Add(time.Hour)))
+	})
+
+	t.Run("NanosecondsTruncated", func(t *testing.T) {
+		startTime := time.Date(2009, time.November, 10, 23, 0, 0, 1000001, time.UTC)
+		allTestIDs := []string{
+			NewTestID(startTime),
+		}
+		tr, err := testExecutionWindow(allTestIDs, string(allTestIDs[0]))
+		assert.NoError(t, err)
+		assert.True(t, tr.StartAt.Equal(time.Date(2009, time.November, 10, 23, 0, 0, 1000000, time.UTC)))
+		assert.True(t, tr.EndAt.Equal(TimeRangeMax))
+	})
+
+	t.Run("NoTestID", func(t *testing.T) {
+		allTestIDs := []string{
+			NewTestID(time.Time{}),
+		}
+		tr, err := testExecutionWindow(allTestIDs, "")
+		assert.NoError(t, err)
+		assert.True(t, tr.StartAt.Equal(TimeRangeMin))
+		assert.True(t, tr.EndAt.Equal(TimeRangeMax))
+	})
+
+	t.Run("NonExistentTestID", func(t *testing.T) {
+		allTestIDs := []string{
+			NewTestID(time.Time{}),
+		}
+		_, err := testExecutionWindow(allTestIDs, "DNE")
+		assert.Error(t, err)
+	})
+}
+
+func TestParseTestIDs(t *testing.T) {
+	for name, testCase := range map[string]struct {
+		keys          []string
+		expectedIDs   []string
+		errorExpected bool
+	}{
+		"EmptyList": {
+			keys:        []string{},
+			expectedIDs: []string{},
+		},
+		"NoMetadata": {
+			keys:        []string{"key1", "key2"},
+			expectedIDs: []string{},
+		},
+		"MalformedMetadata": {
+			keys:          []string{"asdfgh/tests/0de0b6b3Bf4ac6400000000000000000/metadata.json"},
+			errorExpected: true,
+		},
+		"MetadataAndLogChunk": {
+			keys: []string{
+				"builds/asdfgh/tests/0de0b6b3Bf4ac6400000000000000000/metadata.json",
+				"builds/asdfgh/tests/0de0b6b3Bf4ac6400000000000000000/1000000000301000000_1000000000302000000_2",
+			},
+			expectedIDs: []string{"0de0b6b3Bf4ac6400000000000000000"},
+		},
+		"Sorted": {
+			keys: []string{
+				"builds/asdfgh/tests/0de0b6b3cb3688400000000000000000/metadata.json",
+				"builds/asdfgh/tests/0de0b6b3Bf4ac6400000000000000000/metadata.json",
+			},
+			expectedIDs: []string{"0de0b6b3Bf4ac6400000000000000000", "0de0b6b3cb3688400000000000000000"},
+		},
+	} {
+		t.Run(name, func(t *testing.T) {
+			testIDs, err := parseTestIDs(testCase.keys)
+			if testCase.errorExpected {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
+				assert.ElementsMatch(t, testCase.expectedIDs, testIDs)
 			}
-			assert.NoError(t, db.C(TestsCollection).Find(bson.M{}).One(&testDoc))
-			assert.Equal(t, objectID, testDoc.ID)
 		})
-
-		t.Run("TestIDInsertedAsString", func(t *testing.T) {
-			require.NoError(t, testutil.ClearCollections(TestsCollection))
-			test := Test{Id: NewTestID(time.Time{})}
-			require.NoError(t, db.C(TestsCollection).Insert(test))
-
-			var testDoc struct {
-				ID string `bson:"_id"`
-			}
-			assert.NoError(t, db.C(TestsCollection).Find(bson.M{}).One(&testDoc))
-			assert.Equal(t, string(test.Id), testDoc.ID)
-		})
-
-		t.Run("FindObjectID", func(t *testing.T) {
-			require.NoError(t, testutil.ClearCollections(TestsCollection))
-			objectID := bson.NewObjectId()
-			require.NoError(t, db.C(TestsCollection).Insert(bson.M{"_id": objectID}))
-
-			var test Test
-			assert.NoError(t, db.C(TestsCollection).Find(bson.M{"_id": TestID(objectID.Hex())}).One(&test))
-			assert.Equal(t, objectID.Hex(), string(test.Id))
-		})
-
-		t.Run("FindTestID", func(t *testing.T) {
-			require.NoError(t, testutil.ClearCollections(TestsCollection))
-			testID := NewTestID(time.Time{})
-			require.NoError(t, db.C(TestsCollection).Insert(bson.M{"_id": string(testID)}))
-
-			var test Test
-			assert.NoError(t, db.C(TestsCollection).Find(bson.M{"_id": testID}).One(&test))
-			assert.Equal(t, testID, test.Id)
-		})
-	})
+	}
 }
