--- storage/retrieval.go
@@ -1,291 +1,0 @@
-package storage
-
-import (
-	"context"
-	"encoding/json"
-	"sort"
-	"strings"
-	"sync"
-	"time"
-
-	"github.com/evergreen-ci/pail"
-	"github.com/mongodb/grip"
-	"github.com/mongodb/grip/recovery"
-	"github.com/pkg/errors"
-)
-
-// CheckMetadata returns whether the metadata file exists for the given build
-// or test. If the test ID is not empty, the metadata of the test for the given
-// build is checked, otherwise the top-level build metadata is checked. A build
-// ID is required in both cases.
-func (b *Bucket) CheckMetadata(ctx context.Context, buildID string, testID string) (bool, error) {
-	var key string
-	if testID == "" {
-		key = metadataKeyForBuild(buildID)
-	} else {
-		key = metadataKeyForTest(buildID, testID)
-	}
-
-	iter, err := b.List(ctx, key)
-	if err != nil {
-		return false, errors.Wrap(err, "listing metadata files")
-	}
-
-	for iter.Next(ctx) {
-		// We set the prefix to the entire metadata filename, so
-		// finding a single file in the bucket with the given prefix
-		// suffices.
-		return true, nil
-	}
-
-	return false, errors.Wrap(iter.Err(), "iterating metadata files")
-}
-
-// FindBuildByID returns the build metadata for the given ID from the offline
-// blob storage bucket.
-func (b *Bucket) FindBuildByID(ctx context.Context, id string) (*Build, error) {
-	reader, err := b.Get(ctx, metadataKeyForBuild(id))
-	if pail.IsKeyNotFoundError(err) {
-		return nil, nil
-	}
-	if err != nil {
-		return nil, errors.Wrapf(err, "getting build metadata for build '%s'", id)
-	}
-
-	var build *Build
-	if err = json.NewDecoder(reader).Decode(build); err != nil {
-		return nil, errors.Wrapf(err, "parsing build metadata for build '%s'", id)
-	}
-
-	return build, nil
-}
-
-// FindTestByID returns the test metadata for the given build ID and test ID
-// from the offline blob storage bucket.
-func (b *Bucket) FindTestByID(ctx context.Context, buildID string, testID string) (*Test, error) {
-	reader, err := b.Get(ctx, metadataKeyForTest(buildID, testID))
-	if pail.IsKeyNotFoundError(err) {
-		return nil, nil
-	}
-	if err != nil {
-		return nil, errors.Wrapf(err, "getting test metadata for build '%s' and test '%s'", buildID, testID)
-	}
-
-	var test *Test
-	if err = json.NewDecoder(reader).Decode(test); err != nil {
-		return nil, errors.Wrapf(err, "parsing test metadata for build '%s' and test '%s'", buildID, testID)
-	}
-
-	return test, nil
-}
-
-// FindTestsForBuild returns all of the test metadata for the given build ID
-// from the offline blob storage bucket.
-func (b *Bucket) FindTestsForBuild(ctx context.Context, buildID string) ([]Test, error) {
-	iterator, err := b.List(ctx, buildTestsPrefix(buildID))
-	if err != nil {
-		return nil, errors.Wrapf(err, "listing test keys for build '%s'", buildID)
-	}
-
-	testIDs := []string{}
-	for iterator.Next(ctx) {
-		if !strings.HasSuffix(iterator.Item().Name(), metadataFilename) {
-			continue
-		}
-
-		testID, err := testIDFromKey(iterator.Item().Name())
-		if err != nil {
-			return nil, errors.Wrapf(err, "parsing test metadata key for build '%s'", buildID)
-		}
-		testIDs = append(testIDs, testID)
-	}
-
-	var wg sync.WaitGroup
-	catcher := grip.NewBasicCatcher()
-	tests := make([]Test, len(testIDs))
-	for i, id := range testIDs {
-		wg.Add(1)
-		go func(testID string, idx int) {
-			defer recovery.LogStackTraceAndContinue("finding test metadata for build from bucket")
-			defer wg.Done()
-
-			test, err := b.FindTestByID(ctx, buildID, testID)
-			if err != nil {
-				catcher.Add(err)
-				return
-			}
-			tests[idx] = *test
-		}(id, i)
-	}
-	wg.Wait()
-
-	if catcher.HasErrors() {
-		return nil, catcher.Resolve()
-	}
-	return tests, nil
-}
-
-// DownloadLogLines returns log lines for a given build ID and test ID. If the
-// test ID is empty, this will return all logs lines in the build.
-func (b *Bucket) DownloadLogLines(ctx context.Context, buildID string, testID string) (chan *LogLineItem, error) {
-	buildKeys, err := b.getBuildKeys(ctx, buildID)
-	if err != nil {
-		return nil, errors.Wrapf(err, "getting keys for build '%s'", buildID)
-	}
-
-	if len(buildKeys) == 0 {
-		return nil, errors.Errorf("no keys found for build '%s", buildID)
-	}
-
-	buildChunks, testChunks, err := parseLogChunks(buildKeys)
-	if err != nil {
-		return nil, errors.Wrapf(err, "parsing log chunks from keys for build '%s'", buildID)
-	}
-	testChunks = filterLogChunksByTestID(testChunks, testID)
-
-	testIDs, err := parseTestIDs(buildKeys)
-	if err != nil {
-		return nil, errors.Wrapf(err, "parsing test IDs from keys for build '%s'", buildID)
-	}
-	tr, err := testExecutionWindow(testIDs, testID)
-	if err != nil {
-		return nil, errors.Wrapf(err, "getting execution window for test '%s'", testID)
-	}
-
-	// Tests should never be filtered by a time range other than AllTime
-	// since we always want to capture all the lines of either a single
-	// test or all tests.
-	return NewMergingIterator(NewBatchedLogIterator(b, testChunks, 4, AllTime), NewBatchedLogIterator(b, buildChunks, 4, tr)).Stream(ctx), nil
-}
-
-// getBuildKeys returns the all the keys contained within the build prefix.
-func (b *Bucket) getBuildKeys(ctx context.Context, buildID string) ([]string, error) {
-	iter, err := b.List(ctx, buildPrefix(buildID))
-	if err != nil {
-		return nil, errors.Wrapf(err, "listing keys for build '%s'", buildID)
-	}
-
-	var keys []string
-	for iter.Next(ctx) {
-		keys = append(keys, iter.Item().Name())
-	}
-
-	if err := iter.Err(); err != nil {
-		return nil, errors.Wrap(err, "iterating build keys")
-	}
-
-	return keys, nil
-}
-
-// parseLogChunks parses build and test log chunks from the buildKeys that correspond to log chunks
-// and sorts them by start time.
-func parseLogChunks(buildKeys []string) ([]LogChunkInfo, []LogChunkInfo, error) {
-	var buildChunks, testChunks []LogChunkInfo
-	for _, key := range buildKeys {
-		if strings.HasSuffix(key, metadataFilename) {
-			continue
-		}
-
-		var info LogChunkInfo
-		if err := info.fromKey(key); err != nil {
-			return nil, nil, errors.Wrap(err, "getting log chunk info from key name")
-		}
-		if info.TestID != "" {
-			testChunks = append(testChunks, info)
-		} else {
-			buildChunks = append(buildChunks, info)
-		}
-	}
-
-	sortLogChunksByStartTime(buildChunks)
-	sortLogChunksByStartTime(testChunks)
-
-	return buildChunks, testChunks, nil
-}
-
-// parseTestIDs parses test IDs from the buildKeys that correspond to test metadata files
-// and sorts them by creation time.
-func parseTestIDs(buildKeys []string) ([]string, error) {
-	var testIDs []string
-	for _, key := range buildKeys {
-		if !strings.HasSuffix(key, metadataFilename) {
-			continue
-		}
-		if !strings.Contains(key, "/tests/") {
-			continue
-		}
-		testID, err := testIDFromKey(key)
-		if err != nil {
-			return nil, errors.Wrap(err, "getting test ID from metadata key")
-		}
-		testIDs = append(testIDs, testID)
-	}
-
-	sort.Slice(testIDs, func(i, j int) bool {
-		return testIDTimestamp(testIDs[i]).Before(testIDTimestamp(testIDs[j]))
-	})
-
-	return testIDs, nil
-}
-
-// testExecutionWindow returns the time range from the creation of this test to
-// the creation of the next test. If the given test ID is empty, the returned
-// time range is unbounded. If there is no subsequent test then the end time is
-// TimeRangeMax.
-//
-// Tests are expected to be run and, thus, logged sequentially. In cases where
-// they overlap, the test execution window can exclude log lines from the test
-// if the subsequent test begins before the previous test ended. To ensure that
-// we capture all the log lines of a test, we do not filter the test chunks by
-// by this time range. This does mean, though, that the build logs returned
-// with the logs the test may be filtered by a time range shorter than that of
-// the test itself—this behavior is okay since tests are expected to be run
-// serially.
-func testExecutionWindow(allTestIDs []string, testID string) (TimeRange, error) {
-	tr := AllTime
-	if testID == "" {
-		return tr, nil
-	}
-
-	var found bool
-	var testIndex int
-	for i, id := range allTestIDs {
-		if string(id) == testID {
-			found = true
-			testIndex = i
-		}
-	}
-	if !found {
-		return tr, errors.Errorf("test '%s' was not found", testID)
-	}
-
-	tr.StartAt = testIDTimestamp(allTestIDs[testIndex]).Truncate(time.Millisecond)
-
-	if testIndex < len(allTestIDs)-1 {
-		tr.EndAt = testIDTimestamp(allTestIDs[testIndex+1]).Truncate(time.Millisecond)
-	}
-
-	return tr, nil
-}
-
-// filterLogChunksByTestID returns the resulting slice of log chunks after
-// filtering for chunks with the given test ID.
-func filterLogChunksByTestID(chunks []LogChunkInfo, testID string) []LogChunkInfo {
-	if testID == "" {
-		return chunks
-	}
-
-	var filteredChunks []LogChunkInfo
-	for _, chunk := range chunks {
-		if chunk.TestID == testID {
-			filteredChunks = append(filteredChunks, chunk)
-		}
-	}
-	return filteredChunks
-}
-
-func sortLogChunksByStartTime(chunks []LogChunkInfo) {
-	sort.Slice(chunks, func(i, j int) bool {
-		return chunks[i].Start.Before(chunks[j].Start)
-	})
-}
